Move identifier to lib and registration to sveltelib
