+++ ts/components/Item.svelte
@@ -5,7 +5,7 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 <script lang="ts">
     import Detachable from "./Detachable.svelte";
 
-    import type { Register, Registration } from "./registration";
+    import type { Register, Registration } from "../sveltelib/registration";
 
     import { getContext, hasContext } from "svelte";
     import { sectionKey } from "./context-keys";
