+++ ts/components/buttons.ts
@@ -1,7 +1,7 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 import type { Writable } from "svelte/store";
-import type { Registration } from "./registration";
+import type { Registration } from "../sveltelib/registration";
 
 export enum ButtonPosition {
     Standalone,
